refactor: 인코딩 알고리즘의 다형성을 위한 인터페이스 정의
