+++ src/main/java/com/url/shortener/utils/encoding/Base62Util.java
@@ -1,16 +1,16 @@
-package com.url.shortener.utils;
+package com.url.shortener.utils.encoding;
 
-import lombok.AccessLevel;
-import lombok.NoArgsConstructor;
+import org.springframework.stereotype.Component;
 
-@NoArgsConstructor(access = AccessLevel.PRIVATE)
-public final class Base62Util {
+@Component
+public class Base62Util implements EncodingUtil {
 
     private static final String CODE = "0123456789ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz";
     private static final int BASE = 62;
     private static final Long INITIAL_PADDING = 1_000_000L;
 
-    public static String encodeFromNumber(long number) {
+    @Override
+    public String encodeFromNumber(long number) {
         char[] codeCharacters = CODE.toCharArray();
         StringBuilder encodedKey = new StringBuilder();
         number += INITIAL_PADDING;
